requirements.txt
PyPDF2
openai
langchain
tiktoken
python-dotenv
langchain_community
pdfplumber
pymupdf